+++ contracts/bindings/delegationmanager.go
@@ -60,21 +60,21 @@ type ISignatureUtilsSignatureWithExpiry struct {
 	Expiry    *big.Int
 }
 
-//// IStrategyManagerDeprecatedStructQueuedWithdrawal is an auto generated low-level Go binding around an user-defined struct.
-//type IStrategyManagerDeprecatedStructQueuedWithdrawal struct {
-//	Strategies           []common.Address
-//	Shares               []*big.Int
-//	Staker               common.Address
-//	WithdrawerAndNonce   IStrategyManagerDeprecatedStructWithdrawerAndNonce
-//	WithdrawalStartBlock uint32
-//	DelegatedAddress     common.Address
-//}
-//
-//// IStrategyManagerDeprecatedStructWithdrawerAndNonce is an auto generated low-level Go binding around an user-defined struct.
-//type IStrategyManagerDeprecatedStructWithdrawerAndNonce struct {
-//	Withdrawer common.Address
-//	Nonce      *big.Int
-//}
+// IStrategyManagerDeprecatedStructQueuedWithdrawal is an auto generated low-level Go binding around an user-defined struct.
+type IStrategyManagerDeprecatedStructQueuedWithdrawal struct {
+	Strategies           []common.Address
+	Shares               []*big.Int
+	Staker               common.Address
+	WithdrawerAndNonce   IStrategyManagerDeprecatedStructWithdrawerAndNonce
+	WithdrawalStartBlock uint32
+	DelegatedAddress     common.Address
+}
+
+// IStrategyManagerDeprecatedStructWithdrawerAndNonce is an auto generated low-level Go binding around an user-defined struct.
+type IStrategyManagerDeprecatedStructWithdrawerAndNonce struct {
+	Withdrawer common.Address
+	Nonce      *big.Int
+}
 
 // DelegationManagerMetaData contains all meta data concerning the DelegationManager contract.
 var DelegationManagerMetaData = &bind.MetaData{
