+++ contracts/script/eigen/DeployEigenLayer.s.sol
@@ -1,10 +1,8 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity =0.8.12;
 
-import { ERC20 } from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-
+import { MockERC20 } from "test/common/MockERC20.sol";
 import { IStrategy } from "eigenlayer-contracts/src/contracts/interfaces/IStrategy.sol";
-
 import { EigenLayerLocal } from "test/avs/eigen/deploy/EigenLayerLocal.sol";
 import { IEigenDeployer } from "test/avs/eigen/deploy/IEigenDeployer.sol";
 import { Script } from "forge-std/Script.sol";
@@ -38,7 +36,7 @@ contract DeployLocalEigenLayer is Script, EigenLayerLocal {
         for (uint256 i = 0; i < deps.strategies.length; i++) {
             IStrategy strat = IStrategy(deps.strategies[i]);
             strategiesJson =
-                vm.serializeAddress(strategies, ERC20(address(strat.underlyingToken())).symbol(), address(strat));
+                vm.serializeAddress(strategies, MockERC20(address(strat.underlyingToken())).symbol(), address(strat));
         }
 
         // join stragies map with base json
