+++ test/e2e/netman/avs/xdapp.go
@@ -94,8 +94,7 @@ func (d *XDapp) ExportDeployInfo(i types.DeployInfos) {
 	i.Set(d.chain.ID, types.ContractELDelegationManager, d.eigen.DelegationManager, elHeight)
 	i.Set(d.chain.ID, types.ContractELStrategyManager, d.eigen.StrategyManager, elHeight)
 	i.Set(d.chain.ID, types.ContractELPodManager, d.eigen.EigenPodManager, elHeight)
-	i.Set(d.chain.ID, types.ContractELWETH, d.eigen.WETH, elHeight)
-	i.Set(d.chain.ID, types.ContractELWETHStrategy, d.eigen.WETHStrategy, elHeight)
+	i.Set(d.chain.ID, types.ContractELWETHStrategy, d.eigen.Strategies["WETH"], elHeight)
 }
 
 func (d *XDapp) deployOmniAVS(ctx context.Context, client backend.Backend, txOpts *bind.TransactOpts,
